Fix: Change type of ServiceCost, FuelCost, FuelAmount and delete dead code from ApplicationDbContext

+++ Backend/CarServiceApi/CarServiceApi/Models/FuelLog.cs
@@ -7,7 +7,7 @@
         public Vehicle? Vehicle { get; set; }
         public DateTime Date { get; set; }
         public int CarKmCount { get; set; }
-        public int FuelAmount { get; set; } 
-        public int FuelCost { get; set; }
+        public double FuelAmount { get; set; } 
+        public decimal FuelCost { get; set; }
     }
 }
